Move `extensions/pkg/inject` to `pkg/controllerutils/inject`
